@@ -2,6 +2,7 @@
   "name": "hackathon-management-tool",
   "version": "0.1.0",
   "private": true,
+  "type": "module",
   "scripts": {
     "dev": "concurrently --names client,backend --prefix-colors green,#DD00FF --kill-others --passthrough-arguments \"next dev\" \"npx ampx sandbox --profile {@} \" ",
     "dev:client": "next dev",
@@ -20,24 +21,19 @@
     "@aldabil/react-scheduler": "^2.9.5",
     "@aws-amplify/adapter-nextjs": "^1.0.18",
     "@aws-amplify/ui-react": "^6.1.5",
-    "@aws-sdk/client-ses": "^3.598.0",
-    "@emotion/react": "^11.13.3",
-    "@emotion/styled": "^11.13.0",
-    "@react-email/components": "^0.0.19",
-    "@react-email/render": "^0.0.15",
+    "@emotion/react": "11.13.3",
+    "@emotion/styled": "11.13.0",
+    "@mui/icons-material": "^5.15.10",
+    "@mui/material": "^5.15.10",
+    "@mui/x-date-pickers": "^6.19.4",
     "@tanstack/match-sorter-utils": "^8.19.4",
-    "@tanstack/query-async-storage-persister": "^5.36.0",
     "@tanstack/react-query": "^5.32.1",
-    "@tanstack/react-query-devtools": "^5.32.0",
-    "@tanstack/react-query-persist-client": "^5.36.0",
     "@tanstack/react-table": "^8.20.5",
-    "@typescript-eslint/parser": "^7.11.0",
     "@yudiel/react-qr-scanner": "^2.0.4",
-    "aws-amplify": "^6.0.18",
+    "aws-amplify": "^6.15.4",
     "aws-lambda": "^1.0.7",
-    "concurrently": "^8.2.2",
     "contentful": "^10.15.0",
-    "hackathon-management-tool": "file:",
+    "date-fns": "^4.1.0",
     "js-sha3": "^0.9.3",
     "luxon": "^3.4.4",
     "next": "^14.2.12",
@@ -52,30 +48,36 @@
     "validator": "^13.12.0"
   },
   "devDependencies": {
-    "@aws-amplify/backend": "^1.0.0",
-    "@aws-amplify/backend-cli": "^1.0.1",
+    "@aws-amplify/backend": "^1.16.1",
+    "@aws-amplify/backend-cli": "^1.8.0",
     "@aws-sdk/client-cognito-identity-provider": "^3.588.0",
-    "@trivago/prettier-plugin-sort-imports": "^4.3.0",
+    "@ianvs/prettier-plugin-sort-imports": "^4.5.1",
+    "@next/eslint-plugin-next": "^15.4.5",
+    "@tanstack/react-query-devtools": "^5.84.1",
     "@types/aws-lambda": "^8.10.138",
     "@types/luxon": "^3.4.2",
     "@types/node": "^20.12.7",
     "@types/react": "^18",
     "@types/react-dom": "^18",
     "@types/validator": "^13.11.10",
-    "@typescript-eslint/eslint-plugin": "^7.11.0",
+    "@typescript-eslint/eslint-plugin": "^8.38.0",
+    "@typescript-eslint/parser": "^8.38.0",
     "autoprefixer": "^10.0.1",
     "aws-cdk": "^2.126.0",
     "aws-cdk-lib": "^2.126.0",
+    "concurrently": "^9.2.0",
     "constructs": "^10.3.0",
-    "eslint": "^8",
+    "eslint": "^9.32.0",
     "eslint-config-next": "14.0.4",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-prettier": "^5.1.3",
-    "eslint-plugin-promise": "^6.1.1",
-    "eslint-plugin-tailwindcss": "^3.18.0",
+    "eslint-plugin-prettier": "^5.5.3",
+    "eslint-plugin-promise": "^7.2.1",
+    "eslint-plugin-react": "^7.37.5",
+    "eslint-plugin-tailwindcss": "^3.18.2",
     "husky": "^9.1.7",
     "lint-staged": "^15.2.2",
     "postcss": "^8",
+    "prettier": "^3.6.2",
     "prettier-plugin-tailwindcss": "^0.5.11",
     "tailwindcss": "^3.3.0",
     "typescript": "^5.3.3"
