@@ -1,5 +1,5 @@
 /* tslint:disable */
-/* eslint-disable */
+
 //  This file was automatically generated and should not be edited.
 
 export type GenericFunctionResponse = {
