@@ -1,9 +1,8 @@
 import { Amplify } from "aws-amplify";
 import { generateClient } from "aws-amplify/data";
 import type { AppSyncResolverHandler } from "aws-lambda";
-
-import type { Schema } from "../../../data/resource";
 import { data } from "../_amplify_model/amplify_outputs.json";
+import type { Schema } from "../../../data/resource";
 import { updateUser } from "./graphql/mutations";
 
 const MAX_TEAM_MEMBERS = 6;
