@@ -1,7 +1,7 @@
 // /** @type {import("prettier").Config} */
 const config = {
   plugins: [
-    require.resolve("@trivago/prettier-plugin-sort-imports"),
+    require.resolve("@ianvs/prettier-plugin-sort-imports"),
     require.resolve("prettier-plugin-tailwindcss"),
   ],
   endOfLine: "auto",
