@@ -1,12 +1,10 @@
 "use client";
 
 import { useCallback, useMemo, useState } from "react";
-
 import type { Schema } from "@/amplify/data/resource";
-import { client } from "@/app/QueryProvider";
 import { teamColumns } from "@/app/admin/teams/TeamTableSetup";
+import { client } from "@/app/QueryProvider";
 import tanstackTableHelper from "@/components/TanstackTableHelper";
-
 import TableSearch from "./TableSearch";
 import TeamTableBody from "./TanstackTableBody";
 import TableFooter from "./TanstackTableFooter";
