@@ -1,7 +1,6 @@
 "use client";
 
 import { twMerge } from "tailwind-merge";
-
 import { UserType, useUser } from "../contexts/UserContext";
 import Card from "./Card";
 
@@ -30,7 +29,7 @@ export default function Greetings({
         )}
       >
         <h2 className="text-lg font-normal md:text-xl lg:text-2xl">
-          Hack the Change 2024
+          Hack the Change 2025
         </h2>
 
         <h1 className="text-3xl md:text-4xl lg:text-6xl">
