@@ -1,6 +1,5 @@
 import Image from "next/image";
 import Link from "next/link";
-
 import CTC_Icon from "@/CTCLogo.svg";
 
 const arrow_icon = "/svgs/admin/simple_arrow.svg";
@@ -20,91 +19,65 @@ interface DashboardRoutes {
   group: string;
   routes: DashboardLink[];
 }
-const NavLinkContainer = ({
-  dashboardLink,
-}: {
-  dashboardLink: DashboardLink;
-}) => {
-  return (
-    <Link
-      href={dashboardLink.route}
-      className={
-        "mb-2 flex justify-between rounded-md p-2 hover:bg-dark-grey/20"
-      }
-    >
-      <div className="flex gap-2">
-        <div className={"flex justify-center rounded-md bg-white p-2"}>
-          <Image
-            height={20}
-            width={20}
-            src={dashboardLink.icon}
-            alt={`${dashboardLink.name} icon`}
-          />
-        </div>
-        <p className="flex items-center">{dashboardLink.name}</p>
-      </div>
-      <Image height={10} width={10} src={arrow_icon} alt="Arrow icon" />
-    </Link>
-  );
-};
-export default function SideNavBar() {
-  const dashboardRoutes: DashboardRoutes[] = [
-    {
-      group: "ADMIN",
-      routes: [{ name: "Dashboard", icon: dashboard_icon, route: "/admin" }],
-    },
-    {
-      group: "PARTICIPANTS",
-      routes: [
-        {
-          name: "View All Participants",
-          icon: users_icon,
-          route: "/admin/users",
-        },
-        {
-          name: "View All Teams",
-          icon: team_icon,
-          route: "/admin/teams",
-        },
-      ],
-    },
-    {
-      group: "FOOD TICKETS",
-      routes: [
-        {
-          name: "Scan Food Tickets",
-          icon: ticket_icon,
-          route: "/admin/scan-food-tickets",
-        },
-        {
-          name: "Create Food Event",
-          icon: add_icon,
-          route: "/admin/create-food-event",
-        },
-      ],
-    },
-    {
-      group: "RESET",
-      routes: [
-        {
-          name: "Reset Hackathon",
-          icon: reset_icon,
-          route: "/admin/reset",
-        },
-      ],
-    },
-    {
-      group: "SCHEDULE",
-      routes: [
-        {
-          name: "Judging Schedule",
-          icon: reset_icon,
-          route: "/admin/schedule",
-        },
-      ],
-    },
-  ];
 
+const dashboardRoutes: DashboardRoutes[] = [
+  {
+    group: "ADMIN",
+    routes: [{ name: "Dashboard", icon: dashboard_icon, route: "/admin" }],
+  },
+  {
+    group: "PARTICIPANTS",
+    routes: [
+      {
+        name: "View All Participants",
+        icon: users_icon,
+        route: "/admin/users",
+      },
+      {
+        name: "View All Teams",
+        icon: team_icon,
+        route: "/admin/teams",
+      },
+    ],
+  },
+  {
+    group: "FOOD TICKETS",
+    routes: [
+      {
+        name: "Scan Food Tickets",
+        icon: ticket_icon,
+        route: "/admin/scan-food-tickets",
+      },
+      {
+        name: "Create Food Event",
+        icon: add_icon,
+        route: "/admin/create-food-event",
+      },
+    ],
+  },
+  {
+    group: "RESET",
+    routes: [
+      {
+        name: "Reset Hackathon",
+        icon: reset_icon,
+        route: "/admin/reset",
+      },
+    ],
+  },
+  {
+    group: "SCHEDULE",
+    routes: [
+      {
+        name: "Judging Schedule",
+        icon: reset_icon,
+        route: "/admin/schedule",
+      },
+    ],
+  },
+];
+
+export default function SideNavBar() {
   return (
     <details
       className={`transition-width md:open:w-80" relative z-10 flex h-full w-20 cursor-pointer flex-col items-center gap-2 bg-awesomer-purple text-white duration-500 open:fixed open:top-0 open:w-60 md:w-[80px]`}
@@ -135,3 +108,31 @@ export default function SideNavBar() {
     </details>
   );
 }
+
+const NavLinkContainer = ({
+  dashboardLink,
+}: {
+  dashboardLink: DashboardLink;
+}) => {
+  return (
+    <Link
+      href={dashboardLink.route}
+      className={
+        "mb-2 flex justify-between rounded-md p-2 hover:bg-dark-grey/20"
+      }
+    >
+      <div className="flex gap-2">
+        <div className={"flex justify-center rounded-md bg-white p-2"}>
+          <Image
+            height={20}
+            width={20}
+            src={dashboardLink.icon}
+            alt={`${dashboardLink.name} icon`}
+          />
+        </div>
+        <p className="flex items-center">{dashboardLink.name}</p>
+      </div>
+      <Image height={10} width={10} src={arrow_icon} alt="Arrow icon" />
+    </Link>
+  );
+};
